pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mma-model"
version = "0.1.0"
description = "UFC stats ETL, evolving composites, calibrated win probs, EV vs market odds"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "httpx>=0.27",
  "beautifulsoup4>=4.12",
  "lxml>=5.0",
  "sqlalchemy>=2.0",
  "pydantic-settings>=2.0",
  "python-dotenv>=1.0",
  "numpy>=1.26",
  "scikit-learn>=1.4",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "ruff>=0.6",
]

[project.scripts]
mma-model = "mma_model.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
